Server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Kulikov Aleksandr <alexx324@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
django-environ = "^0.12.0"
flake8 = "^7.1.2"
uvicorn = "^0.34.0"
psycopg = "^3.2.6"
djangorestframework = "^3.15.2"
pillow = "^11.1.0"
databases = "^0.9.0"
asyncpg = "^0.30.0"
asgiref = "^3.8.1"
aiohttp = "^3.11.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
